pyproject.toml
[project]
name = "signal-laundering-deepfake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.10.*",
    "numpy==2.4.*",
    "scipy==1.17.*",
    "pandas==3.0.*",
    "matplotlib==3.10.*",
    "seaborn==0.13.*",
    "soundfile==0.13.*",
    "tensorboard==2.20.*",
    "torchcontrib==0.0.2",
    "tqdm==4.67.*",
    "pyyaml==6.0.*",
    "transformers==4.38.*",
    "pyroomacoustics==0.10.*",
]
